use int instead size_t

+++ src/tesseract.cpp
@@ -112,13 +112,13 @@ Rcpp::List engine_info_internal(TessPtr ptr){
   GenericVector<STRING> langs;
   api->GetAvailableLanguagesAsVector(&langs);
   Rcpp::CharacterVector available = Rcpp::CharacterVector::create();
-  size_t I = langs.size();
-  for (size_t i = 0; i < I; i++)
+  int I = langs.size();
+  for (int i = 0; i < I; i++)
     available.push_back(langs.getorat(i).c_str());
   langs.clear();
   api->GetLoadedLanguagesAsVector(&langs);
   Rcpp::CharacterVector loaded = Rcpp::CharacterVector::create();
-  for(size_t i = 0; i < I; i++)
+  for(int i = 0; i < I; i++)
     loaded.push_back(langs.getorat(i).c_str());
   return Rcpp::List::create(
     Rcpp::_["datapath"] = api->GetDatapath(),
